+++ backend/DAL/Subplatform/Survey/Questions/QuestionRepository.cs
@@ -16,8 +16,14 @@ public class QuestionRepository : IQuestionRepository
     public Question ReadQuestionById(int questionId)
     {
         return _dbContext.Questions
-            .FirstOrDefault(q => q.Id == questionId)
-            ?? throw new KeyNotFoundException($"Question with id {questionId} not found.");
+            .SingleOrDefault(q => q.Id == questionId);
+    }
+
+    public Question ReadQuestionByIdWithProject(int questionId)
+    {
+        return _dbContext.Questions
+            .Include(q => q.Project)
+            .SingleOrDefault(q => q.Id == questionId);
     }
 
     public IReadOnlyCollection<Question> ReadAllQuestions()
@@ -25,9 +31,24 @@ public class QuestionRepository : IQuestionRepository
         return _dbContext.Questions.ToList().AsReadOnly();
     }
 
+    public IReadOnlyCollection<Question> ReadAllQuestionsWithProject()
+    {
+        return _dbContext.Questions
+            .Include(q => q.Project)
+            .ToList().AsReadOnly();
+    }
+
     public IReadOnlyCollection<Question> ReadQuestionsByProjectId(int projectId)
     {
         return _dbContext.Questions
+            .Where(q => q.Project.Id == projectId)
+            .ToList().AsReadOnly();
+    }
+
+    public IReadOnlyCollection<Question> ReadQuestionsByProjectIdWithProject(int projectId)
+    {
+        return _dbContext.Questions
+            .Include(q => q.Project)
             .Where(q => q.Project.Id == projectId)
             .ToList().AsReadOnly();
     }
@@ -44,25 +65,73 @@ public class QuestionRepository : IQuestionRepository
         _dbContext.SaveChanges();
     }
 
-    public void DeleteQuestion(int questionId)
+    public bool DeleteQuestion(int questionId)
     {
-        var question = _dbContext.Questions.Find(questionId)
-            ?? throw new KeyNotFoundException($"Question with id {questionId} not found.");
+        var question = _dbContext.Questions
+            .SingleOrDefault(q => q.Id == questionId);
+        if (question == null) return false;
+
         _dbContext.Questions.Remove(question);
         _dbContext.SaveChanges();
+        return true;
     }
 
     public TextAnswer ReadTextAnswerById(int answerId)
     {
         return _dbContext.TextAnswers
+            .SingleOrDefault(a => a.Id == answerId);
+    }
+
+    public TextAnswer ReadTextAnswerByIdWithYouth(int answerId)
+    {
+        return _dbContext.TextAnswers
             .Include(a => a.Youth)
-            .FirstOrDefault(a => a.Id == answerId)
-            ?? throw new KeyNotFoundException($"TextAnswer with id {answerId} not found.");
+            .SingleOrDefault(a => a.Id == answerId);
+    }
+
+    public TextAnswer ReadTextAnswerByIdWithQuestion(int answerId)
+    {
+        return _dbContext.TextAnswers
+            .Include(a => a.Question)
+            .SingleOrDefault(a => a.Id == answerId);
+    }
+
+    public TextAnswer ReadTextAnswerByIdWithYouthAndQuestion(int answerId)
+    {
+        return _dbContext.TextAnswers
+            .Include(a => a.Youth)
+            .Include(a => a.Question)
+            .SingleOrDefault(a => a.Id == answerId);
     }
 
     public IReadOnlyCollection<TextAnswer> ReadTextAnswersByQuestionId(int questionId)
     {
         return _dbContext.TextAnswers
+            .Where(a => a.Question.Id == questionId)
+            .ToList().AsReadOnly();
+    }
+
+    public IReadOnlyCollection<TextAnswer> ReadTextAnswersByQuestionIdWithYouth(int questionId)
+    {
+        return _dbContext.TextAnswers
+            .Include(a => a.Youth)
+            .Where(a => a.Question.Id == questionId)
+            .ToList().AsReadOnly();
+    }
+
+    public IReadOnlyCollection<TextAnswer> ReadTextAnswersByQuestionIdWithQuestion(int questionId)
+    {
+        return _dbContext.TextAnswers
+            .Include(a => a.Question)
+            .Where(a => a.Question.Id == questionId)
+            .ToList().AsReadOnly();
+    }
+
+    public IReadOnlyCollection<TextAnswer> ReadTextAnswersByQuestionIdWithYouthAndQuestion(int questionId)
+    {
+        return _dbContext.TextAnswers
+            .Include(a => a.Youth)
+            .Include(a => a.Question)
             .Where(a => a.Question.Id == questionId)
             .ToList().AsReadOnly();
     }
@@ -79,25 +148,73 @@ public class QuestionRepository : IQuestionRepository
         _dbContext.SaveChanges();
     }
 
-    public void DeleteTextAnswer(int answerId)
+    public bool DeleteTextAnswer(int answerId)
     {
-        var answer = _dbContext.TextAnswers.Find(answerId)
-            ?? throw new KeyNotFoundException($"TextAnswer with id {answerId} not found.");
+        var answer = _dbContext.TextAnswers
+            .SingleOrDefault(a => a.Id == answerId);
+        if (answer == null) return false;
+
         _dbContext.TextAnswers.Remove(answer);
         _dbContext.SaveChanges();
+        return true;
     }
 
     public IntegerAnswer ReadIntegerAnswerById(int answerId)
     {
         return _dbContext.IntegerAnswers
+            .SingleOrDefault(a => a.Id == answerId);
+    }
+
+    public IntegerAnswer ReadIntegerAnswerByIdWithYouth(int answerId)
+    {
+        return _dbContext.IntegerAnswers
             .Include(a => a.Youth)
-            .FirstOrDefault(a => a.Id == answerId)
-            ?? throw new KeyNotFoundException($"IntegerAnswer with id {answerId} not found.");
+            .SingleOrDefault(a => a.Id == answerId);
+    }
+
+    public IntegerAnswer ReadIntegerAnswerByIdWithQuestion(int answerId)
+    {
+        return _dbContext.IntegerAnswers
+            .Include(a => a.Question)
+            .SingleOrDefault(a => a.Id == answerId);
+    }
+
+    public IntegerAnswer ReadIntegerAnswerByIdWithYouthAndQuestion(int answerId)
+    {
+        return _dbContext.IntegerAnswers
+            .Include(a => a.Youth)
+            .Include(a => a.Question)
+            .SingleOrDefault(a => a.Id == answerId);
     }
 
     public IReadOnlyCollection<IntegerAnswer> ReadIntegerAnswersByQuestionId(int questionId)
     {
         return _dbContext.IntegerAnswers
+            .Where(a => a.Question.Id == questionId)
+            .ToList().AsReadOnly();
+    }
+
+    public IReadOnlyCollection<IntegerAnswer> ReadIntegerAnswersByQuestionIdWithYouth(int questionId)
+    {
+        return _dbContext.IntegerAnswers
+            .Include(a => a.Youth)
+            .Where(a => a.Question.Id == questionId)
+            .ToList().AsReadOnly();
+    }
+
+    public IReadOnlyCollection<IntegerAnswer> ReadIntegerAnswersByQuestionIdWithQuestion(int questionId)
+    {
+        return _dbContext.IntegerAnswers
+            .Include(a => a.Question)
+            .Where(a => a.Question.Id == questionId)
+            .ToList().AsReadOnly();
+    }
+
+    public IReadOnlyCollection<IntegerAnswer> ReadIntegerAnswersByQuestionIdWithYouthAndQuestion(int questionId)
+    {
+        return _dbContext.IntegerAnswers
+            .Include(a => a.Youth)
+            .Include(a => a.Question)
             .Where(a => a.Question.Id == questionId)
             .ToList().AsReadOnly();
     }
@@ -114,11 +231,14 @@ public class QuestionRepository : IQuestionRepository
         _dbContext.SaveChanges();
     }
 
-    public void DeleteIntegerAnswer(int answerId)
+    public bool DeleteIntegerAnswer(int answerId)
     {
-        var answer = _dbContext.IntegerAnswers.Find(answerId)
-            ?? throw new KeyNotFoundException($"IntegerAnswer with id {answerId} not found.");
+        var answer = _dbContext.IntegerAnswers
+            .SingleOrDefault(a => a.Id == answerId);
+        if (answer == null) return false;
+
         _dbContext.IntegerAnswers.Remove(answer);
         _dbContext.SaveChanges();
+        return true;
     }
 }
